Fix log, log2, log10 for small arguments (fixes #30)
The different log implementations erroneously used `::epsilon()` to
check if the input argument was "roughly zero". However, especially
in case of the log function, you definitely want to be able to pass
arguments down to `std::numeric_limits<>::min()`.

This change replaces the check by a simple check against zero and
adds test cases to confirm the implementation works correctly down
to the smallest representable number.

# Conflicts:
#	include/gcem_incl/log.hpp
#	include/gcem_incl/log10.hpp
#	include/gcem_incl/log2.hpp
#	tests/log.cpp
#	tests/log2.cpp